refactor code tree
