+++ Iterex/Entity/Tile/Tile.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Collections.Generic;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+using Microsoft.Xna.Framework.Input;
+
+namespace Iterex.Entity.Tile
+{
+    public class Tile
+    {
+        public int id;
+
+        public Tile(int idIn)
+        {
+            id = idIn;
+        }
+
+        public void Draw()
+        {
+
+        }
+    }
+}
