+++ Iterex/Iterex.csproj
@@ -42,14 +42,14 @@
     <ApplicationManifest>app.manifest</ApplicationManifest>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Entity.cs" />
+    <Compile Include="Entity\Entity.cs" />
     <Compile Include="Game1.cs" />
     <Compile Include="Global.cs" />
-    <Compile Include="Player.cs" />
+    <Compile Include="Entity\Player\Player.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Tile.cs" />
-    <Compile Include="World.cs" />
+    <Compile Include="Entity\Tile\Tile.cs" />
+    <Compile Include="Entity\Tile\World.cs" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="MonoGame.Framework">
@@ -109,6 +109,11 @@
       <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
     </None>
     <None Include="app.manifest" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Entity\Creep\" />
+    <Folder Include="Entity\Player\Item\" />
+    <Folder Include="Entity\Player\Skill\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(MSBuildExtensionsPath)\MonoGame\v3.0\MonoGame.Content.Builder.targets" />
