--- Iterex/Tile.cs
@@ -1,23 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Input;
-
-namespace Iterex
-{
-    public class Tile
-    {
-        public int id;
-
-        public Tile(int idIn)
-        {
-            id = idIn;
-        }
-
-        public void Draw()
-        {
-
-        }
-    }
-}
